Cleanup and fixes for Windows sockets
On Windows the `TCPConnection` code was not immediately unsubscribing to ASIO events when trying to shut down a connection. It was also not ever checking to see if all the relevant IOCP operations were complete, so in certain cases it would wait forever on the ASIO event and hang the program.  This change makes `TCPConnection` immediately unsubscribe. There may be some pending IOCP events left over but since we're shutting down we don't care about them reaching client code.

On Windows the recommended way to check the status of a connection is to use `SO_CONNECT_TIME` rather than `SO_ERROR`. Certain failing IOCP operations will not set `SO_ERROR`, so this changes `TCPClient._is_sock_connected()` to use `SO_CONNECT_TIME` on Windows.

Fixes a case in `TCPConnection._event_notify` where clients might receive spurious `connecting` calls when a connection in a different network family failed.

Changes the use of `Sleep()` on Windows to `SleepEX()`, which will wake up for IO operations.

Adds a new test `net/TCPConnectionFailed` that tests if a connection to a bogus port will fail immediately. This was broken on Windows.

Adds a new test for Unix only `new/TCPConnectionToClosedServerFailed` that tests if a connection to a recently-closed server will fail immediately. This test is disabled on Windows because it seems that per investigation of #3656 listening sockets that have an `AcceptEx` in process stick around in the `CLOSE_WAIT` state and will allow connections even if the accepting and listening sockets have been closed.